+++ fltk/src/printer.rs
@@ -5,19 +5,23 @@ use std::ffi::CString;
 /// Defines a printer object
 /// Example usage:
 /// ```no_run
-/// let mut printer = printer::Printer::default();
-/// if printer.begin_job(1).is_ok() {
-///     printer.begin_page();
-///     let (width, height) = printer.printable_rect();
-///     draw::set_draw_color(Color::Black);
-///     draw::set_line_style(draw::LineStyle::Solid, 2);
-///     draw::draw_rect(0, 0, width, height);
-///     draw::set_font(Font::Courier, 12);
-///     printer.set_origin(width / 2, height / 2);
-///     printer.print_widget(widget, -widget.width() / 2, -widget.height() / 2);
-///     printer.end_page();
-///     printer.end_job();
-/// }
+/// use fltk::*;
+/// let mut but = button::Button::default();
+/// but.set_callback2(|widget| {
+///     let mut printer = printer::Printer::default();
+///     if printer.begin_job(1).is_ok() {
+///         printer.begin_page();
+///         let (width, height) = printer.printable_rect();
+///         draw::set_draw_color(Color::Black);
+///         draw::set_line_style(draw::LineStyle::Solid, 2);
+///         draw::draw_rect(0, 0, width, height);
+///         draw::set_font(Font::Courier, 12);
+///         printer.set_origin(width / 2, height / 2);
+///         printer.print_widget(widget, -widget.width() / 2, -widget.height() / 2);
+///         printer.end_page();
+///         printer.end_job();
+///     }
+/// });
 /// ```
 pub struct Printer {
     _inner: *mut Fl_Printer,